% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome.R
\docType{package}
\name{coda4microbiome}
\alias{coda4microbiome}
\title{coda4microbiome: Compositional Data Analysis for Microbiome Studies}
\description{
This package provides a set of functions to explore and study microbiome data within the CoDA framework,
with a special focus on identification of microbial signatures (variable selection).
}
