% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_functions.R
\name{coda_glmnet0}
\alias{coda_glmnet0}
\title{coda_glmnet0}
\usage{
coda_glmnet0(
  x,
  lrX,
  idlrX,
  nameslrX,
  y,
  covar = NULL,
  lambda = "lambda.1se",
  alpha = 0.9
)
}
\arguments{
\item{x}{.}

\item{lrX}{.}

\item{idlrX}{.}

\item{nameslrX}{.}

\item{y}{.}

\item{covar}{.}

\item{lambda}{.}

\item{alpha}{.}
}
\value{
.
}
\description{
Internal function for the permutational test
}
\author{
M. Calle - T. Susin
}
