% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_longitudinal_functions.R
\name{coda_glmnet_longitudinal0}
\alias{coda_glmnet_longitudinal0}
\title{coda_glmnet_longitudinal0}
\usage{
coda_glmnet_longitudinal0(
  x,
  lrX,
  idlrX,
  nameslrX,
  y,
  x_time,
  subject_id,
  ini_time,
  end_time,
  covar = NULL,
  ktop = NULL,
  lambda = "lambda.1se",
  alpha = 0.9,
  nfolds = 10
)
}
\arguments{
\item{x}{abundance table in long format (several rows per individual)}

\item{lrX}{log-ratio matrix}

\item{idlrX}{indices table in the log-ratio matrix}

\item{nameslrX}{colnames of the log-ratio matrix}

\item{y}{outcome (binary)}

\item{x_time}{observation times}

\item{subject_id}{subject id}

\item{ini_time}{initial time to be analyzed}

\item{end_time}{end time to be analyzed}

\item{covar}{data frame with covariates (default = NULL)}

\item{ktop}{given number of selected taxa or compute the best number in case it is NULL (default = NULL)}

\item{lambda}{penalization parameter (default = "lambda.1se")}

\item{alpha}{elastic net parameter (default = 0.9)}

\item{nfolds}{number of folds}
}
\value{
.
}
\description{
internal function
}
\author{
M. Calle - T. Susin
}
