% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_data.R
\docType{data}
\name{Crohn}
\alias{Crohn}
\title{Crohn}
\format{
The dataset contains two objects:
\describe{
\item{x_Crohn}{microbiome abundance matrix for 975 individuals (rows) and 48 genera (columns)}
\item{y_Crohn}{a \code{factor}, indicating if the sample corresponds to a case (\emph{CD}) or a control (\emph{no}).}
}
}
\description{
Microbiome composition at genus level from a Crohn's disease study: 48 taxa and 975 individuals
(662 patients with Crohn`s disease and 313 controls)
}
\references{
\doi{10.1016/j.chom.2014.02.005}
}
\keyword{data}
