% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_functions.R
\name{impute_zeros}
\alias{impute_zeros}
\title{impute_zeros}
\usage{
impute_zeros(x)
}
\arguments{
\item{x}{abundance table (rows are samples, columns are variables (taxa))}
}
\value{
x abundance table with zeros substituted by imputed positive values
}
\description{
Simple imputation: When the abundance table contains zeros, a positive value is added
to all the values in the table. It adds 1 when the minimum of table is larger than 1 (i.e. tables of counts) or
it adds half of the minimum value of the table, otherwise.
}
\examples{

data(HIV, package = "coda4microbiome")

x<-impute_zeros(x_HIV)

}
\author{
M. Calle - T. Susin
}
