% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_longitudinal_functions.R
\name{integralFun}
\alias{integralFun}
\title{integralFun}
\usage{
integralFun(x, y, id, a, b)
}
\arguments{
\item{x}{abundance matrix or data frame in long format (several rows per individual)}

\item{y}{outcome (binary); data type: numeric, character or factor vector}

\item{id}{subjects-ids}

\item{a}{interval initial time}

\item{b}{interval final time}
}
\value{
matrix with integrals for each individual (rows) and each taxa (columns)
}
\description{
Integral of the curve trajectory of subject_id in the interval a,b
}
\author{
M. Calle - T. Susin
}
