% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_functions.R
\name{plot_signature}
\alias{plot_signature}
\title{plot_signature}
\usage{
plot_signature(vars, coeff, showPlots = TRUE)
}
\arguments{
\item{vars}{variables selected}

\item{coeff}{associated coefficients}

\item{showPlots}{if TRUE, shows the plots (default = TRUE)}
}
\value{
bar plot
}
\description{
Graphical representation of the variables selected and their coefficients
}
\examples{

plot_signature(c(2,10, 3, 15, 4), c(0.8, -0.1, 0.2, -0.6, -0.3))


}
\author{
M. Calle - T. Susin
}
