% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_longitudinal_functions.R
\name{plotMedianCurve}
\alias{plotMedianCurve}
\title{plotMedianCurve}
\usage{
plotMedianCurve(iNum, iDen, X, Y, x_time, subject_id, ini_time, end_time)
}
\arguments{
\item{iNum}{.}

\item{iDen}{.}

\item{X}{.}

\item{Y}{.}

\item{x_time}{.}

\item{subject_id}{.}

\item{ini_time}{.}

\item{end_time}{.}
}
\value{
.
}
\description{
Internal plot function
}
\author{
M. Calle - T. Susin
}
