% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_functions.R
\name{shannon}
\alias{shannon}
\title{shannon}
\usage{
shannon(x)
}
\arguments{
\item{x}{abundance composition (vector)}
}
\value{
shannon information
}
\description{
Shannon information
}
\examples{

data(HIV, package = "coda4microbiome")

shannon(x_HIV[1,])


}
\author{
M. Calle - T. Susin
}
