% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_functions.R
\name{shannon_effnum}
\alias{shannon_effnum}
\title{shannon_effnum}
\usage{
shannon_effnum(x)
}
\arguments{
\item{x}{abundance composition (vector)}
}
\value{
shannon information
}
\description{
Shannon effective number of variables in a composition
}
\examples{

data(HIV, package = "coda4microbiome")

shannon_effnum(x_HIV[1,])

}
\author{
M. Calle - T. Susin
}
