% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_functions.R
\name{shannon_sim}
\alias{shannon_sim}
\title{shannon_sim}
\usage{
shannon_sim(x, y)
}
\arguments{
\item{x}{abundance composition (vector)}

\item{y}{abundance composition (vector)}
}
\value{
shannon similarity (value between 0 and 1)
}
\description{
Shannon similarity between two compositions
}
\examples{

data(HIV, package = "coda4microbiome")

shannon_sim(x_HIV[1,],x_HIV[2,])

}
\author{
M. Calle - T. Susin
}
