% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_data.R
\docType{data}
\name{x_ecam}
\alias{x_ecam}
\title{ecam_filtered}
\format{
The dataset contains three objects:
\describe{
\item{x_ecam}{microbiome abundance matrix in long format (873 rows) and 37 genera (columns)}
\item{taxanames}{vector containing the taxonomy of the 37 taxa}
\item{metadata}{matrix with information on the individuals at the observation time}
}
}
\description{
Microbiome composition at genus level from Early childhood and the microbiome (ECAM) study (Bokulich et al. 2016).
Metadata and microbiome data were downloaded from https://github.com/caporaso-lab/longitudinal-notebooks.
Filtered data contains information on 42 children and 37 taxa.
}
\references{
Bokulich et al. (2016). Antibiotics, birth mode, and diet shape microbiome maturation during early life. Sci Transl Med 8:343
}
\keyword{data}
