% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_data.R
\docType{data}
\name{x_HIV}
\alias{x_HIV}
\title{HIV}
\format{
The dataset contains three objects:
\describe{
\item{x_HIV}{microbiome abundance matrix for 155 individuals (rows) and 60 genera (columns)}
\item{y_HIV}{a factor, specifying if the individual is HIV positive or (\code{Pos}) or negative (\code{Neg}).}
\item{MSM_HIV}{a factor, indicating sexual preferences: \code{MSM} (\emph{Men who have Sex with Men}) or not (\code{nonMSM}).}
}
}
\description{
Microbiome abundances (60 taxa and 155 individuals) from an HIV study (Noguera-Julian et al. 2016).
}
\references{
\doi{10.1016/j.ebiom.2016.01.032}
}
\keyword{data}
