% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda4microbiome_data.R
\docType{data}
\name{x_sCD14}
\alias{x_sCD14}
\title{sCD14}
\format{
The dataset contains two objects:
\describe{
\item{x_sCD14}{microbiome abundance matrix for 151 individuals (rows) and 60 genera (columns)}
\item{y_sCD14}{a \code{numeric} variable with the value of the inflammation parameter sCD14 for each sample}
}
}
\description{
Microbiome composition (60 taxa and 151 individuals) and inflammatory parameter sCD14
from an HIV study (Noguera-Julian et al. 2016).
A dataset containing the number of counts of 60 different genera in a group
of 151 samples (including HIV - infected and non - infected patients).
}
\references{
Rivera-Pinto et al. (2018) Balances: a new perspective for microbiome analysis. mSystems 3 (4)
}
\keyword{data}
